bq tag normalization
This change introduces a length normalized barcode score. The main reason is that downstream analysis tools should be agnostic of the barcode design. The bq tag in the BAM stays as int32, following the contract that it may never exceed the range of int8 to support the PBI bq tag, which is int8.

Necessary changes:
 - PPA: normalize bq score in BarcodeCaller::SingleAlign()
 - PPA: use a new HotStartLogitScore()
 - Inform downstream users that the score is now normalized