Update wallStart wallEnd language
In production bams, only an approximate time estimate is available. As of this
writing, the estimate should be accurate to within a minute, but that may
change without notice.

The original language also specifed that wallEnd would be the startFrame of the
base corresponding to qe. Query ranges are end-exclusive, however, and the
startframe of the last base in the read is what was originally sold to
stakeholders. Additionally, with this change the final wallEnd for a ZMW will
be a base location, and not the movie length in frames.